+++ new-portfolio/src/components/common/navbar/Navbar.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import NavLinks from "./navbar/NavLinks";
-import "../../assets/css/nav-bar.css"
+import NavLinks from "./NavLinks";
+import "../../../assets/css/nav-bar.css"
 
 function Navbar() {
   const [navbarOpen, setNavbarOpen] = React.useState(false);
